New CLI * Removed seperate experiment parser file - for now merged into main * Generator now also takes an experiment xml file - from that figure out DT and model name
